pipeline/requirements.txt
# Pin unsloth version — diverging releases break GRPO patches
unsloth>=2026.4.4
trl>=0.12.0
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
accelerate>=0.29.0
torch>=2.1.0
numpy>=1.26.0
pyyaml>=6.0
scipy>=1.12.0
